Added Basic classes, some necessary variables
Variables and classes are set up for later functionality, currently
there is an Enemy class, a BasicZombie class to the Enemy class, and a
Player class. Next step is finding out how to use the rectangle method
to do collision detection, which will allow for the building of the
methods in Enemy which calculate the nearest player and set the zombie
pathfinding to move to it. Advanced AI behavior can then be added in the
AI method. Player control will come before advanced AI, however.